# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vibe Shelf API — a community bookshelf of vibe-coded GitHub projects
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  firstName: string;
  lastName: string;
  projectName: string;
  /** @nullable */
  liveLink?: string | null;
  githubLink: string;
  /** @nullable */
  email?: string | null;
  /**
     * Short description of the project
     * @nullable
     */
  description?: string | null;
  /** Hex color code for the book spine */
  spineColor: string;
  /**
     * Week number (1-6)
     * @nullable
     */
  week?: number | null;
  /** AI Builder of the Week badge */
  isBadged?: boolean;
  createdAt: string;
}

export interface BookInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  projectName: string;
  liveLink?: string;
  /** @minLength 1 */
  githubLink: string;
  email?: string;
  description?: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  week?: number;
}

export interface BookUpdate {
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  /** @minLength 1 */
  projectName?: string;
  liveLink?: string;
  /** @minLength 1 */
  githubLink?: string;
  email?: string;
  description?: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  week?: number;
}

export interface ShelfStats {
  totalBooks: number;
  recentBooks: Book[];
}

export interface ErrorResponse {
  error: string;
}